Implement full game logic
After a cell has been filled, all cells in the grid are checked for
the various win conditions. I.e. rows, columns, diagonals. The turn is
then switched to the other player and the next cell clicked will be
filled with their marker.

In the event that the board is completely filled without a valid winner,
a tie is declared.

Currently, the winner is output to the console. This needs to be
updated to provide an appropriate UI/display.